Moved build_type into the new build_info pkg
